split dispatcher_client class
